# specs/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Nutrimate API documentation
  version: 1.0.0
tags:
  - name: food
    description: Endpoints for foods
  - name: recipes
    description: Endpoint for recipe search
  - name: allergens
    description: Endpoints for allergens
  - name: environmentalScore
    description: Endpoint for environmental score
  - name: user
    description: Endpoints for user and auth

paths:
  /fridge/food:
    get:
      tags:
        - food
      summary: Get a list of all the user's food in the fridge
      operationId: getFood
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Food'
        '401':
          description: User not authenticated
        '204':
          description: No food inside the fridge
    post:
      tags:
        - food
      summary: Add a new food to the user's fridge
      operationId: addFood
      requestBody:
        description: Create a new food instance in the fridge
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodCreationObject'
        required: true
      responses:
        '201':
          description: Food saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '400':
          description: Invalid input
        '401':
          description: User not authenticated

  /food/{barcode}:
    get:
      tags:
        - food
      summary: Find food by barcode
      description: Returns a single food by barcode
      operationId: getFoodByBarcode
      parameters:
        - name: barcode
          in: path
          description: barcode of food to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodScanObject'
        '404':
          description: Food not found in the database

  /food/{barcode}/matching-allergens:
    get:
      tags:
        - food
      summary: Get allergens of a food that match the user's allergens
      description: Get a list of allergens of the food that match the user's allergens
      operationId: getMatchingAllergens
      parameters:
        - name: barcode
          in: path
          description: barcode of food to return allergens of
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Allergen'
        '401':
          description: User not authenticated
        '404':
          description: Food not found

  /fridge/food/{foodId}:
    patch:
      tags:
        - food
      summary: Open a food and add days in which it should be consumed
      operationId: openFood
      parameters:
        - name: foodId
          in: path
          description: ID of food to open
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodOpenObject'
      responses:
        '200':
          description: Food successfully updated
        '401':
          description: User not authenticated
        '400':
          description: Invalid input
        '404':
          description: Food not found
    get:
      tags:
        - food
      summary: Find food by ID
      description: Returns a single food by ID
      operationId: getFoodById
      parameters:
        - name: foodId
          in: path
          description: ID of food to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '401':
          description: User not authenticated
        '404':
          description: Food not found
    delete:
      tags:
        - food
      summary: Deletes a food
      description: delete a food from the user's fridge
      operationId: deleteFood
      parameters:
        - name: foodId
          in: path
          description: ID of food to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Food successfully deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Food'
        '401':
          description: User not authenticated
        '404':
          description: Food not found

  /food/{barcode}/environmental-score:
    get:
      tags:
        - food
      summary: Get environmental score of a food
      description: Get an integer describing the environmental score of a single food
      parameters:
        - name: barcode
          in: path
          description: ID of food to show environmental score of
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentalScore'
        '404':
          description: Food not found

  /recipes:
    get:
      tags:
        - recipes
      summary: Returns a filtered list of recipes
      description: Returns a list of recipes containing the foods and not containing the allergens
      operationId: searchRecipes
      parameters:
        - in: query
          name: categories
          schema:
            type: array
            items:
              type: string
          description: The names of the food categories (1 per food) that the recipe should contain
        - in: query
          name: allergens
          schema:
            type: array
            items:
              type: integer
          description: The ids of the allergens of the user that should not be in the recipes
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '204':
          description: No recipes found

  /fridge/environmental-score:
    get:
      tags:
        - environmentalScore
      summary: Get environmental score of fridge
      description: Get an integer describing the environmental score of all the food in the fridge
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentalScore'
        '401':
          description: User not authenticated
        '204':
          description: No food in the fridge

  /allergens:
    get:
      tags:
        - allergens
      summary: Get all allergens
      description: Get a list of all allergens
      operationId: getAllergens
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Allergen'

  /user/allergens:
    get:
      tags:
        - user
      summary: Get all allergens the user has saved in their profile
      description: Get a list of all the user's allergens
      operationId: getAllergensOfUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Allergen'
        '204':
          description: No content
        '401':
          description: User not authenticated
    patch:
      tags:
        - user
      summary: Change the list of allergens in a user's profile
      operationId: editAllergens
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Allergen'
        required: true
      responses:
        '201':
          description: Allergens saved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Allergen'

        '400':
          description: Invalid input
        '204':
          description: User has no allergenes
        '401':
          description: User not authenticated

  /user/favourite-recipes:
    get:
      tags:
        - user
      summary: Get all recipes the user has saved in their favourites
      description: Get a list of all the user's favourite recipes
      operationId: getRecipesOfUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeFavourite'
        '204':
          description: No content
        '401':
          description: User not authenticated
    post:
      tags:
        - user
      summary: Add one favourite recipe to the list of favourite recipes in a user's profile
      operationId: addFavouriteRecipes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
        required: true
      responses:
        '201':
          description: Recipe saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'

        '400':
          description: Invalid input
        '401':
          description: User not authenticated

  /user/favourite-recipes/{recipeId}:
    delete:
      tags:
        - user
      summary: Removes a recipe from the user's favourites
      description: remove a recipe
      operationId: deleteRecipe
      parameters:
        - name: recipeId
          in: path
          description: ID of recipe to remove
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Recipe successfully removed
          content:
              application/json:
                schema:
                    $ref: '#/components/schemas/RecipeFavourite'
        '401':
          description: User not authenticated
        '404':
          description: Recipe not found

  /user:
    delete:
      tags:
        - user
      summary: Delete Account
      description: Deletes the current user's account
      operationId: deleteUser
      responses:
        '200':
          description: Account successfully deleted
        '401':
          description: User not authenticated

  /register:
    post:
      tags:
        - user
      summary: Register a new user
      description: Creates a new user account
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterObject'
        required: true
      responses:
        '201':
          description: Account created successfully
          headers:
            authToken:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: Invalid input

  /login:
    post:
      tags:
        - user
      summary: Logs in a user
      description: Creates a new session token to log in a user
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterObject'
        required: true
      responses:
        '201':
          description: Logged in successfully
          headers:
            authToken:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: Invalid input
        '404':
          description: User not found

  /logout:
    post:
      tags:
        - user
      summary: Logs out a user
      description: Deletes an existing session token to log out a user
      operationId: logoutUser
      responses:
        '200':
          description: Logged out successfully
        '401':
          description: User not authenticated
        '404':
          description: User not found



components:
  schemas:

    EnvironmentalScore:
      type: object
      properties:
        environmentalScore:
          type: integer

    Food:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        barcode:
          type: string
          example: 3045140105502
        expireDate:
          type: string
          format: date
          example: 2024-09-20
        isOpen:
          type: boolean
          example: false
        daysToConsume:
          type: integer
          example: null
        calories:
          type: number
        fat:
          type: number
        saturated fats:
          type: number
        carbs:
          type: number
        sugar:
          type: number
        fibers:
          type: number
        protein:
          type: number
        salt:
          type: number
        allergens:
          type: array
          items:
            $ref: '#/components/schemas/Allergen'
        category:
          type: string
          example: Schokoladenpudding

    FoodCreationObject:
      type: object
      properties:
        expireDate:
          type: string
          format: date
          example: 2024-09-20
        barcode:
          type: string
          example: 3045140105502
        amount:
          type: integer
          example: 1

    FoodScanObject:
      type: object
      properties:
        barcode:
          type: string
          example: 3045140105502
        calories:
          type: number
        fat:
          type: number
        saturated fats:
          type: number
        carbs:
          type: number
        sugar:
          type: number
        fibers:
          type: number
        protein:
          type: number
        salt:
          type: number
        allergens:
          type: array
          items:
            $ref: '#/components/schemas/Allergen'
        category:
          type: string
          example: Schokoladenpudding

    FoodOpenObject:
      type: object
      properties:
        daysToConsume:
          type: integer
          example: 5

    RecipeFavourite:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        url:
          type: string

    Recipe:
      type: object
      properties:
        url:
          type: string

    Allergen:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: lactose

    UserRegisterObject:
      type: object
      properties:
        email:
          type: string
          example: test@example.com

    AuthToken:
      type: string
      example: o435bkjwkfaew